Some cleanup on the user
