+++ backend/graph/model/models_gen.go
@@ -250,6 +250,7 @@ type User struct {
 	Anonymous bool      `json:"anonymous"`
 	BccMember bool      `json:"bccMember"`
 	Audience  *string   `json:"audience"`
+	Email     *string   `json:"email"`
 	Settings  *Settings `json:"settings"`
 	Roles     []string  `json:"roles"`
 }
